+++ docs/data/data-grid/performance/Example.js
@@ -1,0 +1,5 @@
+import { ExampleConcept } from '@mui/x-data-grid-premium';
+
+export default function Example() {
+  return <ExampleConcept />;
+}
