+++ docs/data/data-grid/performance/Example.tsx
@@ -1,0 +1,6 @@
+import { ExampleConcept } from '@mui/x-data-grid-premium'
+
+export default function Example() {
+  console.log(ExampleConcept)
+  return <ExampleConcept />
+}
