+++ docs/data/data-grid/performance/Example.tsx.preview
@@ -1,0 +1,1 @@
+<ExampleConcept />
