+++ docs/pages/_app.js
@@ -19,7 +19,7 @@ import createEmotionCache from 'docs/src/createEmotionCache';
 import findActivePage from 'docs/src/modules/utils/findActivePage';
 import { LicenseInfo } from '@mui/x-license-pro';
 
-import '@mui/x-data-grid/concept/style.css'
+import '@mui/x-data-grid-premium/concept/style.css'
 
 // Remove the license warning from demonstration purposes
 LicenseInfo.setLicenseKey(process.env.NEXT_PUBLIC_MUI_LICENSE);
