+++ packages/grid/x-data-grid/src/index.ts
@@ -16,8 +16,6 @@ export * from './context';
 export * from './colDef';
 export * from './utils';
 
-export { DataGrid as DataGridConcept } from './concept/DataGrid';
-
 export type { DataGridProps, GridExperimentalFeatures } from './models/props/DataGridProps';
 export type { GridToolbarExportProps } from './components/toolbar/GridToolbarExport';
 export type { GridExportFormat, GridExportExtension } from './models/gridExport';
